use a trait to the IoVec instead of forcing a tuple
The code used to assume a (u64, usize) tuple but in order to give more
flexibility to the user, we now take a trait as input. Additionally, the
ReadManyResult stream now also returns that struct to the user (instead
of the read pos alone). This means that the user can pipe additional
context about the read requests and get them back later on when the
result of reads are ready.