apply clippy suggestions on all targets (doc, examples, benches, tests)
So far we don't run clippy on non-lib targets in CI. This commit applies
all clippy suggestions on those so we can start expecting clippied code
everywhere.

I left those changes in their own commit because I'm okay with dropping
them. There are pros and cons to doing this. Not running clippy
everywhere implicitly creates two languages for glommio: a clippy
version and a non-clippy one. This makes the sources harder to read and
harder to review, IMO.