use an Ubuntu 20.04 VM instance instead of a docker container in CI
CircleCI docker instances run on a 3.13 kernel, this commit changes our
CircleCI executor type from `container` to `machine` so that we can
access newer kernels and sudo privileges. The newest kernel available
is 5.4 from Ubuntu 20.04 but once a newer image is made available,
enabling unit testing on CI should be trivial.

This only applies for the build step however, the other ones still use
docker as they are faster this way.

This commit also uses the official rust orb to install rust and we no
longer ask for a specific version of Rust. We use the `stable` channel
instead, except for the `fmt` job that still requires the nightly one.