don't mutate read buffers
One obstacle we face in sharing buffers around is that the consumers
mutate them in various ways. The main use case for this is offsetting
and trimming data when dealing with unaligned reads.

For instance, the user may ask for (0, 256) but since that's not aligned
the system will instead perform a (0, 512) read. In this case, we need
to slice the resulting buffer so that the final user may consume the
result as if the read was performed exactly as requested.

To circumvent this, this commit leverages the preexisting logic in
`ReadResult` in a way that doesn't need a mutable buffer. In other
places such as stream structs, we substitute buffer references to more
generic slices of bytes that we slice accordingly.