return `IoBuffer` instead of `DmaBuffer` and `Vec<u8>`
The current `Source` struct expects consumers to know the type of IO
being performed when retrieving the buffer after a read operation was
completed. When Sources have a single consumer this is a fair behavior
but we plan to mix and match them between tasks. To make things simpler,
we return the `IoBuffer` that wraps either a `DmaBuffer` or a `Vec<u8>`
directly.

This commit also implements `Deref` for `IoBuffer` such that consumers
can access the buffer bytes without needing to know exactly what kind of
buffer they are dealing with.

This will be useful later on because it will make it easy to feed a
buffered IO request with a previous DMA IO buffer, and vice-versa.