fix UB in buffer allocation
https://doc.rust-lang.org/nightly/alloc/alloc/trait.GlobalAlloc.html#safety-1

It is UB to invoke a Layout with non-zero size, and we don't explicitly
test for it. We should.

Those are I/O buffers and a 0-sized I/O buffer makes no sense. So the
behavior here is to fail the allocation.

Thanks to user Nugine for spotting this.

Fixes #351