enable multiple `Source` awaiters
Currently our `Source` struct keeps track of a single waker function.
This commit changes this so that multiple `Task` can register a `Waker`
object on a `Source`. The overhead is kept to a minimum and we perform
aggressive optimization based on the fact that most `Sources` have a
single consumer.