introduce the `IoScheduler`
We are about to share sources between tasks and we need a place to store
inflight sources to do so. There are a few requirements:
* `Sources` must live exactly as long as needed;
* Sharing a `Source` should extend its lifetime accordingly;
* `Sources` should be stored per file;
* Hard and soft linked files should share buffers;

This commit introduces an `IoScheduler` object that lives in each
executor's reactor. The `IoScheduler` is essentially a tree and contains
two kinds of entities:
* `ScheduledFiles` are intermediary nodes in the scheduler tree. They
  represent a single file and they index inflight `ScheduledSources`.
* `ScheduledSources` are pairs of iovecs and sources that are cloneable.

Each `ScheduledFile` is identified by an identity function using the
device and inode id returned by `statx()`. Therefore, links don't count
when it comes to sharing.

`ScheduledSources` are stored in the scheduler by offset to make lookups
efficient.

`ScheduledFiles` and `ScheduledSources` are intrusive members of the
tree structure: they remove themselves from the scheduler as soon as
they go out of scope. That way this doesn't need the cooperation of
applications.