allow to opt-in to io read deduplication
Add a (private) `attach_scheduler` method on `DmaFile` to fetch the
FileScheduler for a given file and enable read deduplication. A
`ScheduledFile` contains the necessary state and logic to share sources
between tasks for a file.

Because this deduplication is unsafe for read-write files, it is only
enabled for `ImmutableFiles`.