fix file stream close method panic
The close method of a file stream assumes its own the underlying file
and panics if it doesn't. Our API explicitly allows a file to be used in
many streams at once (hence the internal Rc). This commit changes this
behavior. When a stream is closed it will attempt to close the
underlying file but will not panic if it fails to do so.