do not open files with O_DIRECT: add it later instead
Inspired by what seastar does.
I thought we could get away with it using retries, but unfortunately
this doesn't play well with O_EXCL: the file is created anyway, even if
open actually fails because of lack of O_DIRECT.

We can always retry later without O_EXCL but that breaks the atomicity
that O_EXCL expects. So let's use the tried and true method of fcntl.