Immutable file
This commit adds an immutable file. The ImmutableFile allow us to code
the user's expectation that in a correct program the contents of this
file will not change. In that case, we will be able to perform
optimizations.

In particular, we will be able to return a read right away if the buffer
representing that read is still alive.

In a normal file this would break read() semantics, because technically
another write could have changed the contents of the buffer in the mean
time.