record deduplicated IO stats separately
IOs are still recorded as they were before but subsequent IOs feeding
from a preexisting buffer account for different counters. This is useful
to keep track of how effective IO deduplication is.

As for other stats, those new counters are kept track of at a task queue
and ring granularity.