Fixes #377: Await Local executor handle in flaky test
The original test's code did not await both joinhandles, but that is not safe or
correct. Unless you await the joinhandle of a local executor, you cannot be 100%
sure that every object owned by a future inside the executor will be dropped
correctly. This could be addressed in the future but is currently outside of the
scope of this.