simplify schedule reference counting.
There are a couple of places in which we are trying to be way too smart
about schedule reference counts: because we will also decrement it for
some reason (because we are dropping a waker or the task), we don't
increment the ref count at that point. To allow for such optimization,
schedule() itself cannot bump the reference count, which has to be left
to the caller.

I get it, atomics are expensive. But unlike the crate that inspired us
with this task code, our executors are always local. There is contention
on channel wake up, that's true, but that should not be nearly as bad as
contention on every operation.

Even if it were, given that we already had a bad bug in this area, I
say simplicity wins.

Our user Dmitry recently opened an issue (379) with one such bug. That
bug started with a double schedule, but even after that was fixed we
still crashed. And that's essentially because we didn't bump the
refcount when called through wake_by_ref.

The code should be a lot more resilient now, more future proof, and
if there are a couple of extra cycles from a relaxed atomic, so be it.

Fixes #379