Begins fixing #179: Add 'tracing' dependency + example tracepoints
Added the very popular 'tracing' library, as well as a basic subscriber for use
in tests, benchmarks, etc. Only includes a few tracepoints as an example and
doesn't begin to pretend it's instrumented the entire system (FAR from it). I
think with the dependency included people will be more inclined to help out
adding tracepoints. Also fixed compiler errors due to inability to infer types,
potentially caused by the addition of this new crate and it's traits, or were
part of a new compiler behavior. All it does is make them more verbose and
precise about the type so it's included in this commit as well. Only tests are
affected by this.