try to run tasks right away
The current behavior is that, when we create a task, we move it to the
back of a queue. However, a better behavior is to try to execute it
right away. Especially for short lived tasks, that means that we don't
need to go back to the queue at all, which should yield better
performance.

In the task switching microbenchmark, performance of spawning tasks
almost doubles with this change.