remove alpha wording from Readme.
We are getting very close to a point in which I am comfortable moving
Glommio out of alpha.

In particular, the limitations mentioned are either things that we fixed
already (NVMe only, lack of async truncate) or we are unlikely to anyway
(like a memory allocator, which should frankly be a separate crate) and
I/O scheduler (we have something basic now for I/O dedup, but it is
unclear we need more than that).