do not reallocate memory when processing shared channels
The implementation used to iterate over the tree of wakers and would
recreate a new tree with the remaining ones. This is not efficient as it
leads to a high number of allocations. Instead, iterate over the
existing tree and reuse the memory in place.

Also, use the same tree to store wakers and checker functions to reduce
the number of lookups and simplify the lifecycle of both.

Additionally, use a `SmallVec` to store wakers and optimize for the case
where a single waker is consuming the shared channel.