update to rust 1.55 and apply Clippy suggestions
Time flies, we ended up skipping 1.54 this time.

On a side node, we need to be more careful because some code was not
passing Clippy even in 1.53. We seem to have an issue whereby
contributors need to have a CircleCI account for it to detect their PR
and run their code.

This is problematic, and we should find a solution, otherwise we can
except more broken/unformatted/undocumented code to end up being merged.

Code that doesn't pass clippy isn't the end of the world as I can fix it
later on. Undocumented code though, that's another story!