move `local` and `local_into` out of `Task<T>` to simplify API discovery
Spawning tasks via a member function of `Task<T>` is not great from an
API discovery point of view but also mean we often need to use the
starfish syntax such as `Task::<()>::local(...)`. A free function makes
it simpler for the type system to infer the return type of the `Task` to
create.

On a side node, it aligns us with tokio.