move all runtime-related function out of `Task<T>`
Functions such as `create_task_queue` intuitively relate to the executor
and not to a `Task`. To make things more easy, we define a free function
`executor()` that is now included in the prelude. It returns a proxy
struct to the `LocalExecutor`. This is where all executor functions
accessible to the user will be located moving forward.

Once Rust gets its act together wrt thread-local-storage we will remove
this proxy entirely and return a direct reference to the `LocalExecutor`
itself. Alas, we can't do that yet.