buffer sources in the `read_many` internal stream
The `read_many` stream now eagerly consumes the input stream until
said stream yields or until we reach 128 in-flight sources. This allows
a maximum of 128 sources to run concurrently. After that, the stream
will wait until the first source is fulfilled.

128 is a constant for now and was chosen to match the depth of our
io_uring submission queues. Although this value is a constant right now,
a further improvement would be to make it dynamic to maximize
throughput and lower IO latencies on the fly.

Baby step though, one thing at a time.