fix failing tests asserting IO errors
Don't assert the strongly typed errors, just the string representations.
The `Io:ErrorKind` of the errors we were matching against used to be
`Other` and have moved to be `Uncategorized`.
The documentation explicitly advises against matching for
`Uncategorized` errors as they are meant to eventually get their own
dedicated `Io:ErrorKind` in future rust versions.