embed the number of executors to create directly in the Placement policy
Instead of having two orthogonal configurations (the number of executors
to create and the CPU placement policy) that can conflict with each
other (for instance, if we provide a custom placement and a different
number of shards to create), merge the two.

A `Placement` now takes (when appropriate) the number of executors to
create in addition to the `CpuSet` to use. The
`LocalExecutorPoolBuilder` constructor now takes a `Placement` directly
instead of the number of executors to create.