align the `LocalExecutorBuilder` API with the pooled version.
The pooled version is presently superior to the non-pooled one
because it uses the `PoolPlacement` logic. With the existing
`LocalExecutorBuilder,` an executor is either pinned to a specific CPU
or none at all. Therefore, it's possible to spawn a single executor
that's fenced to a collection of CPUs using the pooled API but not with
the non-pooled one.

This commit creates a new `Placement` enum that contains placement
strategies that make sense for a single executor. In turn,
`LocalExecutorBuilder::new` now takes a placement strategy, and the
`pin_to_cpu` API is removed.