fix `DmaStreamReader` repeated poll panic
The `DmaStreamReader` used to panic upon registering a waker if one was
already there on the premise that the stream should have a single
consumer. While it is correct that a unique consumer should poll a
`DmaStreamReader,` there are valid scenarios in which we want to replace
an existing `Waker` for a newer one.

One such example is when the user polls many streams in parallel using a
facility like `JoinAll.` `JoinAll` will create one future that polls
all its child futures every time. Therefore, if one of the child
streams is woken up by the reactor, all the streams in the `JoinAll`
will be polled. In the existing code, this would panic, and this commit
fixes it.