fast `ReadResult` deref
The current `Deref` implementation of `ReadResult` is slow as it
traverses the underlying `ScheduledSource,` `Source,` and performs a
conditional jump depending on the type of buffer (IO buffer, heap, etc.)

This is all pretty expensive! Since a `ReadResult` always carries the
`ScheduledSource` with it, we can materialize the pointer to the
underlying memory and return it directly upon demand.