remove `LICENCE-3rdparty.csv` in favor of an on-demand script
Keeping this file in sync is a significant burden on contributors, and
we don't want to have CI update it either because that would imply a lot
of additional, spurious commits.

Since this file is easily generated, we will now provide a script that
produces the correct output on-demand.

CI will still enforce that all our dependencies are using acceptable
licenses.