update dependencies to avoid depending on the `chrono` crate
The `chrono` crate has a security vulnerability, and the ecosystem is
moving away from it, so should we (also it annoys me that `cargo-check`
doesn't pass anymore because of this).

See https://rustsec.org/advisories/RUSTSEC-2020-0159 for details.

Cargo-check can't be configured to tolerate vulnerabilities in dev
dependencies right now. It's a feature request, so maybe it'll come one
day:
https://github.com/EmbarkStudios/cargo-deny/issues/322