add a way to measure the post-reactor scheduler IO latency
Add the plumbings necessary to measure the post-reactor scheduler delay,
i.e., the time it takes for the scheduler to invoke the task that
consumes the result of a fulfilled source. I suspect a bug in Glommio
creates very high latency spikes; this commit is part of my effort to
find it.

There are three kinds of latency measurements I would like Glommio to
have:
* Pre-reactor delay: the delay between the moment an IO is scheduled and
  the moment it enters the ring;
* IO latency: the time a source spends in the kernel;
* Post-reactor delay: the delay this commit measures, as explained
  above.