update doc for previously blocking synchronous operations
`io_uring` doesn't support all the regular file operations yet. In the
past, we were filling the gaps by issuing blocking syscalls directly,
blocking the reactor thread. Because this kind of sucks, we added a
warning in the docs to warn application writers.

Nowadays, we perform these blocking operations in a background thread.
Therefore, we no longer block the reactor thread, and the warning is no
longer relevant. Instead, we replace the warning with a note to let
users know that, depending on the system they run on, the syscall may be
issued in a background thread to achieve asynchronicity.