soft fail when crossing buffer boundaries in `get_buffer_aligned`
The `get_buffer_aligned` API could be better. It relies on user code to
be aware of buffer boundaries locations. This makes user code more
complicated and brittle because the buffer size will generally be a
configurable knob, subject to change.

Fear not, however, for there is a better way!

Instead of failing when reading too much in `get_buffer_aligned,` read
whatever we can from the current buffer. This allows the following:
* If the read fits inside a buffer, the user can consume the result
  directly without any copy;
* If the read crosses buffer boundaries and the user can consume partial
  results, it can do so in a loop without any copies;
* If the read crosses buffer boundaries and the user needs a complete
  result, `get_buffer_aligned` may be used trivially in a loop that
  concatenates the results in a reusable user-allocated buffer.

The following commit will replace the `AsyncRead::poll_read`
implementation with a simple loop calling `get_buffer_aligned.`