submit latency-sensitive IO requests on the latency ring
This fixes an issue whereby a task queue waiting for external events
(those coming from the reactor, like IO or timers) would not be woken up
 according to its latency requirement.

Consider the following setup: we have two task queues, `(A)` and `(B)`.
`(A)` is latency-sensitive with a preemption timer of 1ms and performs
IO. `(B)` doesn't have a latency requirement and therefore defaults to
the executor's timer: 100ms. `(B)` is CPU intensive but highly
cooperative.

And now the following sequence of events:
* `(A)` and `(B)` are both `READY`; therefore, the preemption timer is
  set to the minimum of the two, 1ms;
* `(A)` runs and schedules all the IO it can. The task returns
  `Poll::Pending`;
* `(B)` runs and eventually gets preempted after ~1ms;
* The reactor enters the rings and posts the IO from `(A)`;
* `(A)` is `PENDING` and `(B)` is `READY,` so the preemption timer is
  set to `(B)`: 100ms;
* `(B)` runs for 100ms, until it is preempted;
* The reactor enters the rings and collects the IO results, waking up
  `(A)` in the process;
* `(A)` and `(B)` are both `READY`; therefore, the preemption timer is
  set to the minimum of the two, 1ms;
* `(A)` collects the result of the IOs
* ...

This isn't right: `(A)` wants to be woken up within 1ms of being marked
`READY`! To fix this issue durably, all IO from latency-sensitive queues
are now posted on the latency ring. This means that as soon as the
kernel creates a CQE on the latency ring, tasks get preempted by the
next call to `yield_if_needed.`

This improves the latency of these IO operations significantly. Right
now, this means the poll ring is only used when IO is submitted from
a throughput task queue. We may try to improve upon this later on.