consume the latency ring before reevaluating the preemption timer
Today, we compute the next preemption timer before touching the rings
to avoid entering the rings multiple times. However, this causes latency
regressions because if some CQEs are ready in the latency ring, that
could wake up some tasks, which in turns could influence the next
preemption timer.

We poll the latency ring before recomputing the preemption timer to
avoid undue overhead. It's okay to skip the ones like the IO they carry
belong to throughput queues.