consume the poll ring before reevaluating the preemption timer
Today, we compute the next preemption timer before touching the rings
to avoid entering the rings multiple times. However, this causes latency
regressions because if some CQEs are ready in the poll ring, that could
wake up some tasks, which in turns could influence the next preemption
timer.

We poll the poll ring before recomputing the preemption timer to avoid
undue overhead.