allow polling single-device md devices (#488)
Linux is weird. LVM devices can present themselves as md devices with a
single underlying device. That device may be polled (but ones with
more than a single underlying device can't). Right now, if a
device is md, we opt out of polling entirely, even though polling is
allowed under that edge-case.

This commit changes the logic to disable polling only if an md device
has more than a single device underneath it.