guard against undefined `minimum_io_size`
The wording of the documentation uses `MIGHT` in the way it describes
how device define `minimum_io_size`. In case it returns `0`, we use
`logical_block_size` as a fallback (whichever is larger).

We also use `512` as a minimum alignment to guard against cases where
they both return `0`. That shouldn't happen, but who knows:
Linux is weird.