faster source cancellation
So far, to cancel a source in glommio, the code would:
* Probe the submitting queue of the associated ring to find its
  corresponding `UringOpDescriptor`;
* If one is found, it is removed;
* Otherwise, a cancellation request is sent on the ring.

The issue with this way of handling cancellation is that the first step
is a linear probing over a potentially large number of descriptors.
Specifically, it means that canceling a source is _O(n)_ where _n_ is
the total number of queued sources in the ring.

Therefore, a realistic application canceling a query (that issues
many IOs) is an _O(kn)_ operation where _k_ is the number of sources
created by the request. Uh oh. That's a lot of work done in a destructor
where we can't cooperate and yield.

Instead, this commit changes the behavior of source cancellation to be
lazy. When dropped, a source marks itself as "canceled" if it wasn't
submitted yet. Meanwhile, the reactor now filters out canceled sources.
If the source was dispatched, we submit a cancel request like before.

Co-authored-by: Glauber Costa <glauber@chiselstrike.com>
Co-authored-by: Hippolyte Barraud <hippolyte.barraud@datadoghq.com>