release any unprepared SQEs
When we create an SQE iterator, we advance the tail of the ring buffer
regardless of whether the iterator is consumed. Implementing `Drop` is
safer to release whatever we didn't prepare.

This implementation is safe because holding a `SQEs` implies holding a
mutable reference to the ring.

This will be useful in the next commit of this series because we are
about to filter out canceled sources, so the number of SQEs we ask for
and the number of SQEs we prepare may mismatch.