accept a minimum number of events ready as a timeout condition
This is natively supported by `io_uringf`, A `IO_URING_TIMEOUT` can
fire if either of two conditions are fulfilled: (1) a certain amount of
time has elapsed or (2) a certain number of CQEs have been posted on the
CQ.

Right now we only use the former, hard coding the latter to 0 (i.e. it
is not considered).