better tracking of ready SQEs; make sure we only sleep if we can
Instead of keeping a counter of stuff waiting submission, ask the ring
how many SQEs are ready. Incidentally, this fixes an accounting bug
related to sleeping. Before sleeping, we would ++ the
`waiting_for_submission` counter without an associated --. This would
end up with the reactor calling `io_uring_submit` spuriously forever
afterwards.

Additionally, the `can_sleep` logic was not thorough enough, and we
would allow sleeping with unsubmitted SQEs, or uncollected CQEs.
We now also assert that we successfully submitted the linking SQE
before sleeping, for good measure. Reading the man page for
`io_uring_enter` it is not explicitly guaranteed that the kernel will
accept all the SQEs we submit to it.