install the preemption timer as the very last step
If the preemption timer is tiny there is a chance it may come back
before we exit the reactor (we consume the ring just after submitting
it). That's really not ideal because it could lead to a deadlock
afterwards if user code has any kind of endless job. Even if that's the
not case, it would lead to a much poorer IO polling rate than we'd like.