use the latency ring sparingly
The latency ring is a double-edged sword: it is the best way to get
preempted as soon as the result of an action is available, which is
highly desirable in some cases. However, if we submit too many things on
that ring, we deteriorate the performance of the entire system because
we keep yielding all the time, and our IO concurrency suffers as a
result.

It is a challenging balance to strike, and I feel that our current
level of granularity for selecting a ring (the task queue) is not fine
enough.

For now, centralize the ring dispatching logic in one place and
implement the following rules:
 * Disk reads/writes go to the poll ring if possible, or the main ring
   otherwise;
 * Network Rx and connect/accept go the latency ring;
 * Every other request are dispatched to the main ring;

We avoid putting requests that come in high numbers on the latency ring
because the more request we issue there, the less effective it becomes.

This should provide an okay balance: read/write IO requests go to the
main/poll ring as they often come in batches, Rx operations go to the
latency ring and the rest got to the main ring.

Additionally, this commit disables rushing to the latency ring as it
breaks preemption signaling.