provide a way to configure the blocking thread pool placement strategies
Add knobs in the local executor builders to configure how many blocking
threads to create in the pool and where to place them on the system. By
default, an executor spawns a single blocking thread in its pool and
collocates it with its thread.

It could be a good idea to loosen that default placement strategy to
bind the blocking thread to any CPUs on the same package as the executor
thread. This would maximize resource utilization and minimize memory
latency. However, there is no easy way to do that using our existing
Placement APIs.