fix a vruntime accounting bug
The current implementation resets the vruntime of a queue upon
activation based on the last observed vruntime. This is problematic
because it negates CPU shares if a task queue keeps going to sleep
(because it is submitting and waiting for IO, most likely)

This mechanism is there because we want to balance active task
queues exclusively. When a task queue wakes up, it cannot use its
previous vruntime:  that value is no longer meaningful. Instead, we need
to find some new value that represents the vruntime of the entire
system.

A better solution is to use the minimum vruntime out of all the
currently active task queues - 1. This guarantees that the activated
task queue will run first for at least one scheduling cycle.