force submission of important IO requests
So far, we have been pretty lax when submitting stuff in
the rings. We push a request in there, flush, and if it fails, we keep
going.

This works for user requests that can wait until the next visit to the
reactor but can cause problems for those requests we depend on. One such
request is the preemption timer.

This commit adds a force flush logic that only returns if it has
successfully submitted a specific number of events to the kernel and
will even block as a last resort if it has to.

This replaces the old cancellation flushing logic that would poll the
kernel for one million times and panic afterward.

Additionally, this fixes a bug that would leave cancellations in the
queue, unsubmitted, if the number of cancellations exceeds the SQE ring
capacity (which is pretty small at ~128 typically).