add `read_many_unordered`
Sometimes order doesn't matter, and an application can consume the IO
results in random order. This allows the stream to drive more IO
concurrency than the ordered variant.

In a pathological case, the ordered stream may have completed all IO
requests except the next one to yield. In that situation, the IO
concurrency of the stream drops to 1 because we need the first IO to
come back before we can yield anything and make room to schedule more IO
requests.

This may seem like it shouldn't happen but ywo factors play against us:
 * The kernel doesn't order anything and,
 * Devices don't order anything and,
 * Request coalescing may produce requests of varying sizes. Large IO
 requests take much longer to complete.

An unordered stream will yield the IO results as soon as they arrive so
that the IO concurrency can remain maximal at all times.