fix the unit tests covering `spin_before_park`
These tests were passing but weren't measuring what we thought they were
because of a call to `getrusage` with `RUSAGE_THREAD` made from an
external thread.
Additionally, we were only measuring user time. We now also account for
system time. This makes sense because when glommio is spinning, a
significant portion of CPU time is expected to be spent in the kernel
polling IO sources.