do not assume files are pollable
Currently, Glommio panics when reading and writing using the poll
ring on a file that doesn't support iopoll (also confusingly known as
`hipri` and `mq_poll` at the VFS layer).

Unfortunately, some filesystems don't support polling. Even though we
are explicit in the documentation that we only support filesystems that
do (XFS, EXT4); this is not enough:
* LVM and dmcrypt mess with it;
* EXT4 iopoll support is somewhat recent;

So far, our approach has been that:
* MD devices never support iopoll;
* Everything else supports it.

If that's not the case, however, we currently fail hard. Instead, this
commit adds a new step when opening a file to exercise the
iopoll code path and probe whether it fails or not such that we can fall
back on interrupt-based IO if needed.

On a side note here, I have spent a good amount of time trying to find a
way to determine whether iopoll is supported on a given
/sys/block/device or a given FD, but there's nothing.

Additionally, we currently don't assume that if iopoll works on a
device, it will work for every file on that device. I believe this is a
safe assumption to make, though, and in the future, we could cache that
information to avoid the extra overhead every time a file is opened.